introduce styled-components into the mix. will likely want to refactor all css
