+++ package.json
@@ -45,6 +45,7 @@
     "react-hook-form": "^7.50.1",
     "react-markdown": "^9.0.1",
     "react-router-dom": "^6.21.1",
-    "remark-html": "^16.0.1"
+    "remark-html": "^16.0.1",
+    "styled-components": "^6.1.13"
   }
 }
