+++ src/pages/components/Wrapper/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import style from './index.module.css';
 
-export function NavBtn(loc, func = () =>{}) {
+export function NavBtn(loc, func = ()=>{}) {
 
 	if (typeof func !== 'function') {
 		func = () => {}
