+++ src/pages/components/navBarWrapper/wrapper.styled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+ export const NavBarWrapper = styled.div`
+	grid-area: navBar;
+	display: flex;
+	width: 98%;
+	margin: auto;
+`
